Work around a bug with setting the lang, maintain existing options
It seems the official documentation is wrong/broken and that setting
'lang' in the RecaptchaOptions doesn't actually do anything. The work
around is to append a new parameter to the query string of the
script/iframe. See:
http://code.google.com/p/recaptcha/issues/detail?id=133

This also ensures that if RecaptchaOptions already exists and is an
object that we don't clobber it. This allows for you to set a global
theme in your own javascript while still retaining the ability to
customize individiual recaptcha instances.