Update readme -- needed LocaleMiddleware in Django 1.9
Received an error on line 64 of client.py, which does not catch a NoneType response from django.utils.translation.get_language(). The get_language method only seems to return None when translation support is not enabled. Thus, the Django LocaleMiddleware must be in MIDDLEWARE_CLASSES for this application to function properly in Django 1.9.7 (have not tested with other versions of Django).