Add Django 4.2 and Python 3.11 testing (#305)
* Update compatibility test matrix

Remove Django version from Github CI matrix and only create jobs for
every supported Python. **Reasoning:** not all Django versions are
compatible with every Python. Having to update and/or exclude certain
combinations from the Github matrix is a maintenance overhead.

Failures for a specific Django version can be reviewed by
checking the CI logs.

- Add Django 4.2
- Drop Django 4.0
- Drop Django 2.2

- Add Python 3.11
- Drop Python 3.7

* Restore testing against Django main branch

Make it a separate job not run by default when executing tox. The
in-development version of Django can be unstable and cause failures that
are not relevant when contributing.

* Update changelog

Reword existing entries for consistency with older entries.