Mark all V2 captcha widgets as hidden so no label is rendered
For the RecaptchaV2Invisible widget it does not make sense to show a label.
The widget is invisible.

The RecaptchaV2Checkbox already renders its own checkbox with an appropriate
label. It does not make sense to render another, as that label is not
connected to any form input and is an accessibility concern.